requirements.txt
streamlit
faiss-cpu
pdfplumber
python-docx
pandas
sentence-transformers
requests
paddleocr
Pillow
